support for softnpu-to-softnpu links
== Primary Change

This patch adds a `Runner::softnpu_links` method which allows softnpu
ports to be directly connected.

== Other updates/fixes

* node! macro

Add a `node!` macro to reduce repetitiveness when specifying nodes
        let foo = d.node("foo", 4, gb(4));
becomes
        node!(foo, 4, gb(4))

* remove println! usage in favor of logging

* fix `NAME_REGEX`

This regex was matching any string that contains a valid name. So an
invalid name containing a valid name would pass. Added `^`/`$` to the
beginning/ending of the regex to ensure that an entire string must be a
valid name.