add pfexec as a target runner
All of the tests in this repo require pfexec, so just use that here.
(It's nice that pfexec doesn't cause a user prompt, which would be
problematic most likely).

In the future it would be cool to have some filtering of target runner
by test name (maybe I should add that to nextest?)