Clean up SerialCommander for use in `exec`. (#61)
* Clean up SerialCommander for use in `exec`.

There was a problem where every drain operation waited until its timeout
making it take longer than necessary. Additionally, some of the parsing
was less than ideal.

Instead we now pass through the unedited output from the terminal minus
the command we sent, and the prompt. In other words we strip the first
and last lines of the output. This gives us nice terminal colors for our
output as well.

We may want to also set the terminal size for better sized output as
well right after we login. That can come in a follow up.