use propolis_client for instance serial consoles (#269)
Poking at `ws://{addr}/instance/serial` directly means that `serial.rs`
sneaks by outside the view of API versioning. As of propolis#955,
propolis-server expects a version header to be present on API requests,
which we hadn't done here and didn't automatically get because we're not
using a generated client!

Move to `propolis_client::support::InstanceSerialConsoleHelper` and
avoid the whole issue. This is what Nexus uses to follow instance
serials, including using `propolis_client` and following serials across
migrations (not that that's a concern in Falcon today).